backend/requirements.txt
Flask==2.3.2
pyodbc==4.0.40
PyJWT==2.8.0
bcrypt==4.0.1
python-dotenv==1.0.0
